break up foreign key tests
The file was getting too hard to understand with all of the different permutations of table.
